backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.1.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
ultralytics==8.0.196  # YOLOv8
opencv-python==4.8.1.78
numpy==1.26.2
pytesseract==0.3.10
reportlab==4.0.7  # For PDF generation
openpyxl==3.1.2  # For Excel export 